move mascot to lib
