+++ examples/apps/demo2/src/theme.rs
@@ -22,12 +22,8 @@ pub struct KeyBinding {
 }
 
 pub struct Logo {
-    pub rat: Color,
     pub rat_eye: Color,
     pub rat_eye_alt: Color,
-    pub hat: Color,
-    pub term: Color,
-    pub term_border: Color,
 }
 
 pub struct Email {
@@ -79,12 +75,8 @@ pub const THEME: Theme = Theme {
     description: Style::new().fg(LIGHT_GRAY).bg(DARK_BLUE),
     description_title: Style::new().fg(LIGHT_GRAY).add_modifier(Modifier::BOLD),
     logo: Logo {
-        rat: Color::Rgb(204, 204, 204), // color of the website
-        hat: Color::Rgb(255, 255, 255),
         rat_eye: BLACK,
         rat_eye_alt: RED,
-        term: BLACK,
-        term_border: LIGHT_GRAY,
     },
     key_binding: KeyBinding {
         key: Style::new().fg(BLACK).bg(DARK_GRAY),
