+++ ratatui-widgets/src/logo.rs
@@ -1,6 +1,12 @@
 //! The [`RatatuiLogo`] widget renders the Ratatui logo.
 use indoc::indoc;
-use ratatui_core::{buffer::Buffer, layout::Rect, text::Text, widgets::Widget};
+use ratatui_core::{
+    buffer::Buffer,
+    layout::{Margin, Rect},
+    style::Color,
+    text::Text,
+    widgets::Widget,
+};
 
 /// A widget that renders the Ratatui logo
 ///
@@ -165,6 +171,198 @@ impl Size {
             █▀▀▄ ▄▀▀▄▝▜▛▘▄▀▀▄▝▜▛▘█  █ █
             █▀▀▄ █▀▀█ ▐▌ █▀▀█ ▐▌ ▀▄▄▀ █
         "}
+    }
+}
+
+const RATATUI_MASCOT: [&str; 32] = [
+    "               hhh              ",
+    "             hhhhhh             ",
+    "            hhhhhhh             ",
+    "           hhhhhhhh             ",
+    "          hhhhhhhhh             ",
+    "         hhhhhhhhhh             ",
+    "        hhhhhhhhhhhh            ",
+    "        hhhhhhhhhhhhh           ",
+    "        hhhhhhhhhhhhh     ██████",
+    "         hhhhhhhhhhh    ████████",
+    "              hhhhh ███████████ ",
+    "               hhh ██ee████████ ",
+    "                h █████████████ ",
+    "            ████ █████████████  ",
+    "           █████████████████    ",
+    "           ████████████████     ",
+    "           ████████████████     ",
+    "            ███ ██████████      ",
+    "          bb    █████████       ",
+    "         bxxb   █████████       ",
+    "        bxxxxb ██████████       ",
+    "       bxx█xxxb █████████       ",
+    "      bxx██xxxxb ████████       ",
+    "     bxxxxxxxxxxb ██████████    ",
+    "    bxxxxxxxxxxxxb ██████████   ",
+    "   bxxxxxxx██xxxxxb █████████   ",
+    "  bxxxxxxxxx██xxxxxb ████  ███  ",
+    " bxxxxxxxxxxxxxxxxxxb ██   ███  ",
+    "bxxxxxxxxxxxxxxxxxxxxb █   ███  ",
+    "bxxxxxxxxxxxxxxxxxxxxxb   ███   ",
+    " bxxxxxxxxxxxxxxxxxxxxxb ███    ",
+    "  bxxxxxxxxxxxxxxxxxxxxxb █     ",
+];
+
+/// A widget that renders the Ratatui mascot
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
+pub struct RatatuiMascot {
+    colors: MascotColors,
+}
+
+/// The colors used to render the Ratatui mascot
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// The colors used to render the Ratatui mascot
+pub struct MascotColors {
+    /// The color of the rat
+    pub rat: Color,
+    /// The color of the rat's eye
+    pub rat_eye: Color,
+    /// The color of the rat's hat
+    pub hat: Color,
+    /// The color of the terminal
+    pub term: Color,
+    /// The color of the terminal border
+    pub term_border: Color,
+}
+
+impl Default for MascotColors {
+    fn default() -> Self {
+        Self {
+            rat: Color::Rgb(204, 204, 204), // color of the website
+            hat: Color::Rgb(255, 255, 255),
+            rat_eye: Color::Rgb(8, 8, 8),
+            term: Color::Rgb(8, 8, 8),
+            term_border: Color::Rgb(188, 188, 188),
+        }
+    }
+}
+
+use itertools::Itertools; // tuples();
+
+impl RatatuiMascot {
+    /// Create a new Ratatui mascot widget
+    pub const fn new(colors: MascotColors) -> Self {
+        Self { colors }
+    }
+
+    /// Tiny help to set the color of the rat's eye
+    #[must_use]
+    pub const fn set_eye_color(self, rat_eye: Color) -> Self {
+        Self {
+            colors: MascotColors {
+                rat_eye,
+                ..self.colors
+            },
+        }
+    }
+
+    /// Use half block characters to render a logo based on the `RATATUI_LOGO` const.
+    ///
+    /// The logo is rendered in three colors, one for the rat, one for the terminal, and one for the
+    /// rat's eye. The eye color alternates between two colors based on the selected row.
+    pub fn render_mascot(&self, area: Rect, buf: &mut Buffer) {
+        let area = area.inner(Margin {
+            vertical: 0,
+            horizontal: 2,
+        });
+        let rat_color = self.colors.rat;
+        let term_color = self.colors.term;
+        let hat_color = self.colors.hat;
+        let border_color = self.colors.term_border;
+        for (y, (line1, line2)) in RATATUI_MASCOT.iter().tuples().enumerate() {
+            for (x, (ch1, ch2)) in line1.chars().zip(line2.chars()).enumerate() {
+                let x = area.left() + x as u16;
+                let y = area.top() + y as u16;
+                let cell = &mut buf[(x, y)];
+                match (ch1, ch2) {
+                    ('█', '█') => {
+                        cell.set_char('█');
+                        cell.fg = rat_color;
+                        cell.bg = rat_color;
+                    }
+                    ('h', 'h') => {
+                        cell.set_char('█');
+                        cell.fg = hat_color;
+                    }
+                    ('█', ' ') => {
+                        cell.set_char('▀');
+                        cell.fg = rat_color;
+                    }
+                    ('h', ' ') => {
+                        cell.set_char('▀');
+                        cell.fg = hat_color;
+                    }
+                    (' ', '█') => {
+                        cell.set_char('▄');
+                        cell.fg = rat_color;
+                    }
+                    (' ', 'h') => {
+                        cell.set_char('▄');
+                        cell.fg = hat_color;
+                    }
+                    (' ', 'b') => {
+                        cell.set_char('▄');
+                        cell.fg = border_color;
+                    }
+                    ('b', ' ') => {
+                        cell.set_char('▀');
+                        cell.fg = border_color;
+                    }
+                    ('b', 'b') => {
+                        cell.set_char('█');
+                        cell.fg = border_color;
+                    }
+                    ('b', 'x') => {
+                        cell.set_char('▀');
+                        cell.fg = border_color;
+                        cell.bg = term_color;
+                    }
+                    ('x', 'b') => {
+                        cell.set_char('▄');
+                        cell.fg = border_color;
+                        cell.bg = term_color;
+                    }
+                    ('█', 'x') => {
+                        cell.set_char('▀');
+                        cell.fg = rat_color;
+                        cell.bg = term_color;
+                    }
+                    ('x', '█') => {
+                        cell.set_char('▄');
+                        cell.fg = rat_color;
+                        cell.bg = term_color;
+                    }
+                    ('x', 'x') => {
+                        cell.set_char(' ');
+                        cell.fg = term_color;
+                        cell.bg = term_color;
+                    }
+                    ('█', 'e') => {
+                        cell.set_char('▀');
+                        cell.fg = rat_color;
+                        cell.bg = self.colors.rat_eye;
+                    }
+                    ('e', '█') => {
+                        cell.set_char('▄');
+                        cell.fg = rat_color;
+                        cell.bg = self.colors.rat_eye;
+                    }
+                    (_, _) => {}
+                };
+            }
+        }
+    }
+}
+
+impl Widget for RatatuiMascot {
+    fn render(self, area: Rect, buf: &mut Buffer) {
+        self.render_mascot(area, buf);
     }
 }
